Add files via upload
This file contains three files. The code in documentSummaries.py is designed to generate summaries of documents based on topic modelling. It utilizes a custom class, topicSummary, to represent summaries for specific topics, encompassing key terms, their respective weights, and relevant sentences. The document summarization process is managed by the DocumentSumaries class, which uses a topic model to determine the dominant topics in a document set and then generates summaries for these topics. This involves selecting the most relevant terms and sentences for each topic, leveraging functions like innerProduct and cosineSimilarity to measure textual similarity and relevance. These summaries help in quickly grasping the essential content of large text corpora, focusing on the thematic significance conveyed through the most relevant terms and sentences associated with each topic. We use ROUGE metrics to check the accuracy and precision scores